requirements.txt
fastapi
uvicorn
opencv-python
numpy
pillow
jinja2
python-dotenv
fatsecret
rauth
torch
torchvision
ultralytics
python-multipart